LearningAgent: add File structure
This commit adds a file structure for the project. The structure is
based on the MVC pattern. The main file is the script.py file.
This file is the entry point of the program. The script.py file will
instantiate the controller and the controller will control the
Learning Agents movements and behaviour.