LearningAgent: squish
This commit is squished. I cannot be bothered to write a 2000 word essay on what i did here.

Basically i moved a lot of stuff from TController to GController and implemented the tetris logic.

If you get a full row of blocks then all blocks move down until they hit another block or the end
of the screen.
I had to change a lot of code in TController but everything is better now than it was before.

It was not tested much but i can say with 99% certainty that this will be good enough.

Issue: #4 - Status: DONE