LearningAgent: tetrimino collision revised (for rotation)
We move the landing stuff to move down method so tetriminos can only land when
they are moving down.

For the rotation we rotate the tetrimino temporarily and check if it is colliding
with the bounds or another block and then decide if we want to switch it back or not.
This saves us from calculating any rotation for the tetriminos.
Also moving the landing stuff to move down method prevents the tetriminos to
"stick" to the left and right of blocks because you cannot "land" the tetrimino
when moving it right or left now.
It also makes a lot more sense coding wise.

With this the base game is 99% done.

Issue: 2 , Status: DONE