LearningAgent: change observation space and action space
We set the observation space to block grid and the current tetrimino
which is all the agent needs to know.

Added down to action space even if it is not necessary we might see him
try to "speedrun" to get more points.

Also changed the way Agent interacts with the game we now post events to pygame
so the gameController can pick it up (needs testing)