LearningAgent: fix q table and cleaning
I cleaned the code a little (debug prints in tetriminoController)

The new q table looks like this:
	State MoveLeft MoveRight MoveDown Rotate
And we add new rows instead of filling pre initialized ones, which means
we can now have q tables that are very large.

With state as the index it is also easy to look for q values.

Also added a function to save q table to a csv file called
q_table.csv to make it easier to look at the q table.

Issue: 11 , Status: DONE? , Issue 12 , Status: DONE