add prediction of final tetrimino position
We calculate the position of where the tetrimino is going to land.
This gives the agent a better chance to learn from every single move he makes.

Example: He rotates an I piece to horizontal over an L piece (with those
being the only ones on the Grid)
- The total height of the whole stack would increase meaning the reward for this action
  would be negative
- He leaves a lot of space on the bottom rows meaning the reward would be even more
  negative

Before all the actions prior to the one actually placing the tetrimino would only
represent the current state of the block grid, which is not very helpful for learning.

Now he knows exactly how much every action is going to affect the board, or at least he
could learn it.

Might need more testing.